Modify lock logic.
プロセス間同期処理を入れ替え。

+++ src/cuto/master/main.go
@@ -78,7 +78,7 @@ func realMain(args *arguments) int {
 		config.Log.OutputLevel,
 		config.Log.MaxSizeKB,
 		config.Log.MaxGeneration,
-		config.Log.TimeoutSec); err != nil {
+		int64(config.Log.TimeoutSec)); err != nil {
 		console.Display("CTM021E", err)
 		return rc_ERROR
 	}
